requirements/base.txt
Django==4.2.7
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1